Move web/httputils to pkg/httputil and add DeadlineClient to it
