+++ pkg/httputil/util.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package httputils
+package httputil
 
 import (
 	"encoding/json"
