+++ storage/remote/opentsdb/client.go
@@ -28,7 +28,7 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
-	"github.com/prometheus/prometheus/utility"
+	"github.com/prometheus/prometheus/pkg/httputil"
 )
 
 const (
@@ -50,7 +50,7 @@ type Client struct {
 func NewClient(url string, timeout time.Duration) *Client {
 	return &Client{
 		url:        url,
-		httpClient: utility.NewDeadlineClient(timeout),
+		httpClient: httputil.NewDeadlineClient(timeout),
 	}
 }
