+++ web/api/api.go
@@ -20,9 +20,9 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
+	"github.com/prometheus/prometheus/pkg/httputil"
 	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/storage/local"
-	"github.com/prometheus/prometheus/web/httputils"
 )
 
 // MetricsService manages the /api HTTP endpoint.
@@ -35,7 +35,7 @@ type MetricsService struct {
 // RegisterHandler registers the handler for the various endpoints below /api.
 func (msrv *MetricsService) RegisterHandler(pathPrefix string) {
 	handler := func(h func(http.ResponseWriter, *http.Request)) http.Handler {
-		return httputils.CompressionHandler{
+		return httputil.CompressionHandler{
 			Handler: http.HandlerFunc(h),
 		}
 	}
